5.0.0 dev

+++ tx-manager/src/main/java/com/codingapi/txlcn/manager/core/message/ManagerRpcExceptionHandler.java
@@ -55,11 +55,11 @@ public class ManagerRpcExceptionHandler implements RpcExceptionHandler {
     public void handleNotifyUnitMessageException(Object params, Throwable e) {
         // notify unit message error, write txEx
         List paramList = ((List) params);
-        String modName =  rpcClient.getAppName((String) paramList.get(1));
+        String modName = rpcClient.getAppName((String) paramList.get(1));
 
         NotifyUnitParams notifyUnitParams = (NotifyUnitParams) paramList.get(0);
         WriteTxExceptionDTO writeTxExceptionReq = new WriteTxExceptionDTO(notifyUnitParams.getGroupId(),
-                notifyUnitParams.getUnitId(), modName, (short) notifyUnitParams.getState());
+                notifyUnitParams.getUnitId(), modName, notifyUnitParams.getState());
         writeTxExceptionReq.setRegistrar((short) 0);
         compensationService.writeTxException(writeTxExceptionReq);
     }
