+++ tx-manager/src/main/java/com/codingapi/txlcn/manager/db/domain/TxException.java
@@ -52,7 +52,7 @@ public class TxException {
     /**
      * 事务状态
      */
-    private Short transactionState;
+    private Integer transactionState;
 
     /**
      * 上报方 0 TxManager 1 TxClient
